backend/healthManage/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.8.0rc2
distlib==0.3.9
Django==5.1.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
django-oauth-toolkit==3.0.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.16.0
drf-yasg==1.21.8
filelock==3.16.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
kombu==5.4.2
MarkupSafe==3.0.2
msgpack==1.1.0
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.1
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
routing==0.2.0
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
social-media==0.0.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
Werkzeug==3.1.0
WTForms==3.1.2
zope.event==5.0
zope.interface==7.2
